requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
av==10.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.1
boto3==1.26.112
botocore==1.29.112
cachetools==5.3.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.4
contourpy==1.0.7
cryptography==40.0.1
cycler==0.10.0
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
filelock==3.11.0
fire==0.5.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==2.10
ifaddr==0.2.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.3
roboflow==1.0.3
rsa==4.9
s3transfer==0.6.0
sahi==0.11.13
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.21.0
streamlit-webrtc==0.45.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
yolov5==7.0.11
zipp==3.15.0
